mcp-server-python/requirements.txt
httpx~=0.27.2
mcp~=1.1.0
python-dotenv~=1.0.1
py-clob-client~=0.22.0

